Implement Linux ABI in Newlib

+++ newlib-1.18.0/newlib/libc/machine/riscv/machine/_types.h
@@ -3,19 +3,17 @@
 
 #include <machine/_default_types.h>
 
-// these types keep struct stat in sync with x86-64 linux
-
 #define __dev_t_defined
 typedef unsigned long long __dev_t;
 
 #define __uid_t_defined
-typedef unsigned long long __uid_t;
+typedef unsigned int __uid_t;
 
 #define __gid_t_defined
-typedef unsigned long long __gid_t;
+typedef unsigned int __gid_t;
 
 #define __off_t_defined
-typedef unsigned long long _off_t;
+typedef long long _off_t;
 
 #define __fpos_t_defined
 typedef unsigned long long _fpos_t;
@@ -27,9 +25,9 @@ typedef unsigned long long time_t;
 typedef unsigned long long __ino_t;
 
 #define __nlink_t_defined
-typedef unsigned long long __nlink_t;
+typedef unsigned int __nlink_t;
 
 #define __mode_t_defined
-typedef unsigned long long __mode_t;
+typedef unsigned int __mode_t;
 
 #endif
